Add ugly flex hack to make ios safari behave.
I really dont like this, but it seems like Safari and iOS needs flex-basis to enable wrapping flex tags. Since we are using flex tags and overriding the flex attribute many places, this hack made tags wrap correctly. I have not found any big inconsistencies with this afterwards, as it should only affect mobile devices. Safari 11 and iOS 11 are said to fix this, but we should support wrapping on iOS 10+.