Add script lint:pretter
This adds a check that all files are formatted with the prettier version
used in the project. The pre-commit hook can be omitted, and it doesn't
fix files that isn't staged. This will also help when bumping the
prettier version, becuase it ensures that all files have been fixed.

This also runs the prettier check when 'yarn run lint' is executed.