Implement group permission admin
So far it is possible to add permissions. No validation of the
permission is done in the frontend.

Also, the `editGroup` action redirected to `interestgroups`, for some
reason. My guess is that we tried to remove the interest group acitons,
in order to unite all group actions, but seing as we're still using
them, we probably gave up. This is probably for the better, since there
is some special handling of the actions of interest groups.