Add FETCH_PARTIAL aciton, to fetch all memberships of a group
We want to get all memberships of the group we're looking at, and the
API uses pagination. Now we fetch memberships using the `next` from the
payload until there is nothing left. Since we want to add memberships to
the group on the `next` fetches, but not on the initial fetch, we make a
new action type.