Use timezone when formatting date in datepicker
The containers we run the webapp in use "no" timezone, aka. they just
use +00:00. This results in strange behavior when the server side
renderer formats the date/time on the server. When the datepicker is
re-rendered in the browser, it use out timezone (Europe/Oslo). This
results in a wrong date label, and an annoying flash when clicking
around.