Add pretty console output for tty
Use a pretty output instead of json when the server is running in a tty
(valid terminal). Will still output json when piped, written to a file
or just running in headless.

This is a way better approach than just using __DEV__ or NODE_ENV,
becuase it also prints pretty when running a containerized version of
the server with an pseudo-TTY.

$ docker run -t lego-webapp   # pretty
$ docker run lego-webapp      # json
$ node dist/server.js         # pretty
$ node dist/server.js | cat   # json
$ yarn start                  # pretty
$ yarn start | cat            # json