Quick fix
When an article has no image it lost its width, as the image was the
only thing pushing the card out. This just adds a 100% to ensure that
the card is always streched (mostly for staging).

Also scale down the readme text a bit, i went a bit overboard last time
i scaled it up :/