Update more e2e tests to work with cypress & chromium
Fixes some more cypress tests to work with the editor in
chromium. Due to the hacky way we make this work, this commit also adds
a test that only works on firefox for manually testing. This test is
skipped when running on other browsers.

Also fixes an issue where it was hard to focus on an
editorfield in the joblistings editor.