Add a more informative loading component when waiting for a pending registration.
This component will show between a REGISTRATION.REQUEST_REGISTER_SUCCESS and a
SOCKET_REGISTRATION.SUCCESS or FAILURE. The normal loading component
will still show between the registration initition and the request
register success. Note that since there is no coupling between the
registration and events, this logic is only stored in state and we cannot
determine whether to show the component based on the registration alone.