Show event registration and stripe popups for at least 2 seconds
This makes the "we are processing your registration" box only pop up if
the regsitration takes longer than 2 seconds AND shows it for at least 2
seconds. Similarly for the stripe success message, it will be shown for
at least 2 seconds. This makes it possible for the user to actually read
the contents of the messages.