Fix multiple select bug in image upload
It was not possible to select multiple images from the file manager when
uploading images.

Additionally fixes a bug where the selected files are stored in the
Upload components state, as well as the main ImageUpload component,
which lead to all existing files would be duplicated when selecting
files for the n > 1'st time.

Also adds a small preview of the images next to the filename so it's
easier to see what is selected before uploading, and changes the remove
icon to a larger trash can icon.