Fix a bug where event detail could crash
When selecting the follower status, the selector used the event.id
instead of the eventId from the router params. This meant that in some
cases, the user following status is successfully fetched before the
event, and as such `event.id` is undefined and the selector would throw
an error. This just uses the eventId from the url params, so the page
won't crash when this code runs before the event is successfully
fetched.