Ignore R2D2's import sorting commit when using `git blame`
`git blame --ignore-revs-file .git-blame-ignore-revs` will now output a useful git blame. Sadly, the flag must always be passed, so I recommend setting a config for this; `git config blame.ignoreRevsFile .git-blame-ignore-revs`. Note that this needs to be set per-project, as `git blame` will raise an error if it cannot find the file.