Upgrade ionicons to v5.5.2
Icons are now SVGs, instead of being a font. This is clever, as the icon font would require us to bundle several megabytes of font data, even if we didn't use all the icons. Now with this, we only include the icons we actually need to use in our app.

There is now no difference between "ios" and "md" prefixed icons, so that prop is removed.