Introduce a `.git-blame-ignore-revs` file
I religiously use `git blame`, and the recent bulk changes in regards
to our transition to Typescript has made it awful. But this makes
everything better! `git blame --ignore-revs-file .git-blame-ignore-revs`
will now output a useful `git blame`. The flag must always be
passed, so I recommend setting a config for this;
`git config blame.ignoreRevsFile .git-blame-ignore-revs`. Note that
this needs to be set per-project, as `git blame` will raise an error
if it cannot find the file.