Rewrite overview component to functional
Also refactor alot of the component. Duplicate filter statements are
removed and memoized. Individual components are converted to react
components and declarations moved out of the component.

A bunch of types are fixed as well.

The all should make the page a bit more performant and efficient as well
as easier to maintain.