Introduce the Inter font
Inter is a _very_ popular font-family. It gives the webapp a more modern
look by replacing Raleway and Roboto.

Changes were made to make it fit more, but I believe this is something
we'll have to iterate on multiple times to get right. No more uppercase
headers, and the webapp is less gray.

I'm considering introducing a "cooler" secondary font to use for headers,
but that can come later.