Give the text input fields a retouch
They look more modern, and conform better with our styling elsewhere.

A `prefix` prop has been added, which takes in the name of an icon and
turns it into a prefix.
A `removeBorder` props has also been added, and is used in the registration table.

This also fixes a bug where the disabled text fields on dark mode did not look disabled.