Improve /companies page
The cards now employ the `<Card />` component, making them resemble the
rest of the webapp more. Also some other minimal styling changes.
Finally, the whole card is a link, and not just the image, which has been
an annoying bug for a really long time.

`CompaniesPage` is rewritten to a functional component.

`isHoverable` prop is added for cards that work as links.