Include waiting list to summary pool tab
It makes more sense to include everyone in this tab, and it's annoying to
have to check multiple tabs when looking for someone.

People on waiting list are given an opacity to their row as a visual label.
`isMeeting` prop had to be added, because meeting registrations don't
have pools the same way events does.