Fix cypress tests failing for event_visit_spec
Fixes two errors with the comment testing in the event_visit_spec. The
first is not waiting after cypress.get() which caused an error. The second
is the cypress tests sometimes missing the comment form. Due to
asynchronous fetching of data, the user permissions will be loaded after
the rest of the page. This causes the button for sending an announcement
to appear after the page has been loaded. This *sometimes* caused the
cypress test to fail as the comment moved a bit down between selecting
and clicking. This is "solved" by waiting for a little bit after
loading.