Move meeting end when setting start time after end
When creating a new meeting one often have to set the date multiple
times, both for the end time field and start time field. This commit
changes the form to work similar to a lot of other meeting forms in
changing the end time if one moves the start time past the end time.
Also changes the default time to something more probable.