Introduce a `secondary` button type
And make the `ghost` type resemble it more.

The secondary button shares the color of the default text color, and it looks much more sharp with the darker color. It replaces a lot of success buttons, as it doesn't really make sense for submit buttons to be of type success. In my eyes it should be used on submit buttons and other buttons with "significant value" that should pop out more on pages.