Fix misalignment of anonymous attendance status
It was the tooltip that caused the misalignment, so just repositioning
it fixed the bug. A maximum width was also added to prevent tooltips
with a lot of content text from just becoming one very long line.

Some minor additional styling was added by utilizing the button component,
and disabeling it when the user is not authenticated.