Make checkboxes focusable
With the changes, it is possible to tab through checkboxes and select
them.

Turned out to be impossible to do for radio buttons, as essentially according to W3C; a radio button is a group that functions as a single element since it retains only a single value. Tabbing to a radio group will bring you to the first item and then using the arrow keys you navigate within the group.

However, I did add the similar logic for radio buttons in case there is
no default value to the radio button group (e.g. in the company interest
form). Despite this being bad practice.