Refactor how event types and contact statuses
The way these were previously defined and used was a mess, and it still
sort of is. I find this to be an improvement though. This actually
resolves a couple of bugs as well; in the sorting and wrong colors on
event types. The "new" (but really just the correct) colors are actually
possivle to read on dark theme, which has long been a requested fix.