Update the look of the penalty card
The design might not be the best, but in my opinion it now looks more
"thorough" and improved. It had clearly not been given a lot of love in
the past.

Notice the small things such as the timeout on the "Prikken er
registrert!" feedback, or the now included info and link to the relevant
event.