Fix bugs preventing semesters field from being required
Firstly, the initial values when creating a new interest prevented the
error message from being rendered. Why? Not sure.
Then there' this problem with `MultiSelectGroup` (or actually `FormSpy`)
that would somehow get the errors as an array where the first item is
the message in an object format and the second item `undefined`?
Why? Still a mysery, and I suppose there must be a cleaner solution to
this, but I have already spent more hours than I dare to admit and we
need this deployed asap ..