Prevent MazeMap logo from growing out of toggle button
The Image component it previously used gives its content a width of `100%`. This would override the MazeMap's width of `25px`, and cause it to "grow". However, there's no reason for the logo to use the Image component as it's just an image file in our assets, and not something which needs to be loaded.

Also replace the logo with a better svg file.