Introduce "too late" presence option to event attendees
Neither ion-icons or font-awesome had a turtle icon, so I extracted one
from Lucide (which we're planning on migrating to). I'm very open to
design suggestions, as I'm not sure what I feel about this. The padding
feels a bit much, so I might refactor the Icon component - but that will
probably have to wait.